Fix compilation error on wercker

+++ tests/bit_operation/tests/validator.cc
@@ -9,7 +9,7 @@ int main() {
   inf.readSpace();
   string t = inf.readToken();
   inf.readSpace();
-  int64_t n = inf.readLong(MIN_N, MAX_N);
+  int64_t n = inf.readLong(static_cast<long long>(MIN_N), static_cast<long long>(MAX_N));
   inf.readEoln();
   inf.readEof();
   ensure(func == "least_bit" || func == "least_bit_fast" || func == "most_bit" || func == "count_bit");
